fix/#4: refactor idea API

+++ src/main/java/com/nexters/teambuilder/idea/service/IdeaService.java
@@ -5,6 +5,7 @@ import com.nexters.teambuilder.idea.api.dto.IdeaResponse;
 import com.nexters.teambuilder.idea.domain.Idea;
 import com.nexters.teambuilder.idea.domain.IdeaRepository;
 import com.nexters.teambuilder.idea.exception.IdeaNotFoundException;
+import com.nexters.teambuilder.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,8 +17,8 @@ import java.util.stream.Collectors;
 public class IdeaService {
     private final IdeaRepository ideaRepository;
 
-    public IdeaResponse createIdea(IdeaRequest request) {
-        return IdeaResponse.of(ideaRepository.save(Idea.of(request)));
+    public IdeaResponse createIdea(User author, IdeaRequest request) {
+        return IdeaResponse.of(ideaRepository.save(Idea.of(author, request)));
     }
 
     public IdeaResponse getIdea(Integer ideaId) {
@@ -26,9 +27,14 @@ public class IdeaService {
         return IdeaResponse.of(idea);
     }
 
-    public IdeaResponse updateIdea(Integer ideaId, IdeaRequest request) {
+    public IdeaResponse updateIdea(User author, Integer ideaId, IdeaRequest request) {
+
         Idea idea = ideaRepository.findById(ideaId)
                 .orElseThrow(() -> new IdeaNotFoundException(ideaId));
+
+        if(!idea.getAuthor().getId().equals(author.getId())) {
+            throw new IllegalArgumentException("해당 아이디어의 작성자가 아닙니다");
+        }
 
         idea.update(request);
 
@@ -40,9 +46,14 @@ public class IdeaService {
         return ideaList.stream().map(IdeaResponse::of).collect(Collectors.toList());
     }
 
-    public void deleteIdea(Integer ideaId) {
+    public void deleteIdea(User author, Integer ideaId) {
         Idea idea = ideaRepository.findById(ideaId)
                 .orElseThrow(() -> new IdeaNotFoundException(ideaId));
+
+        if(!idea.getAuthor().getId().equals(author.getId())) {
+            throw new IllegalArgumentException("해당 아이디어의 작성자가 아닙니다");
+        }
+
         ideaRepository.delete(idea);
     }
